BUG: Fixed, compiler warnings, added grayscale static
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=5133
git-svn-id: http://svn.slicer.org/Slicer4/trunk@5133 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/VolumeRendering/vtkSlicerFixedPointVolumeRayCastMapper.h
@@ -92,8 +92,8 @@ public:
     vtkGetMacro(ManualInteractive,int);
     vtkSetMacro(ManualInteractive,int);
     vtkBooleanMacro(ManualInteractive,int);
-    vtkGetMacro(ManualInteractiveRate,int);
-    vtkSetMacro(ManualInteractiveRate,int);
+    vtkGetMacro(ManualInteractiveRate,double);
+    vtkSetMacro(ManualInteractiveRate,double);
 
   //ENDSLICERADD
 
@@ -262,7 +262,7 @@ protected:
 
     //SLICERADD
     int ManualInteractive;
-    int ManualInteractiveRate;
+    double ManualInteractiveRate;
   //ENDSLICERADD
